requirements.txt
# GIS Pipeline — зависимости
# Установка: pip install -r requirements.txt

# Геопространственные
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.5.0
fiona>=1.9.0

# Обработка изображений
numpy>=1.24.0
opencv-python-headless>=4.8.0
scipy>=1.11.0

# Визуализация
matplotlib>=3.7.0

# Данные / утилиты
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Опционально: ускорение
# numba>=0.57.0
